Location Update between Players Part 3
Created a new interface ILocationEventUpdate to let networking side (PlayerOperationHandler etc) know which data actually is transferred. Updater is now a generic class allowing all types of IEventUpdate implementing interfaces and classes. Updater now receives an Action taking PlayerPeerBase and a List of T : IEventUpdates. The action may be defined in PlayerOperationHandler packing the data and sending it as event to the peer.