Introduced ITimedEventChannel.
Puts IEventChannels UpdateFrequency property into a new interface (inherting from IEventChannel). Now an explicit difference between events collected before sending and instant sended events is done. As channel creation should be triggered from the networking side and the map-side only knows during run time which channels are used in which way, this does not violate the devision of netlogic and gamelogic. Additionally, IChatChannel is now a child interface of IEventChannel.